+++ modules/apps/knowledge-base/knowledge-base-service/src/main/java/com/liferay/knowledge/base/settings/definition/KBGroupServiceConfigurationBeanDeclaration.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.knowledge.base.web.internal.configuration.definition;
+package com.liferay.knowledge.base.settings.definition;
 
-import com.liferay.knowledge.base.web.internal.configuration.KBDisplayPortletInstanceConfiguration;
+import com.liferay.knowledge.base.configuration.KBGroupServiceConfiguration;
 import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
 
 import org.osgi.service.component.annotations.Component;
@@ -23,12 +23,12 @@ import org.osgi.service.component.annotations.Component;
  * @author Roberto Díaz
  */
 @Component(service = ConfigurationBeanDeclaration.class)
-public class KBDisplayPortletInstanceConfigurationBeanDeclaration
+public class KBGroupServiceConfigurationBeanDeclaration
 	implements ConfigurationBeanDeclaration {
 
 	@Override
 	public Class<?> getConfigurationBeanClass() {
-		return KBDisplayPortletInstanceConfiguration.class;
+		return KBGroupServiceConfiguration.class;
 	}
 
 }
