+++ modules/apps/knowledge-base/knowledge-base-web/src/main/java/com/liferay/knowledge/base/web/internal/settings/definition/KBSectionPortletInstanceConfigurationBeanDeclaration.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.knowledge.base.web.internal.configuration.definition;
+package com.liferay.knowledge.base.web.internal.settings.definition;
 
-import com.liferay.knowledge.base.web.internal.configuration.KBArticlePortletInstanceConfiguration;
+import com.liferay.knowledge.base.web.internal.configuration.KBSectionPortletInstanceConfiguration;
 import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
 
 import org.osgi.service.component.annotations.Component;
@@ -23,12 +23,12 @@ import org.osgi.service.component.annotations.Component;
  * @author Roberto Díaz
  */
 @Component(service = ConfigurationBeanDeclaration.class)
-public class KBArticlePortletInstanceConfigurationBeanDeclaration
+public class KBSectionPortletInstanceConfigurationBeanDeclaration
 	implements ConfigurationBeanDeclaration {
 
 	@Override
 	public Class<?> getConfigurationBeanClass() {
-		return KBArticlePortletInstanceConfiguration.class;
+		return KBSectionPortletInstanceConfiguration.class;
 	}
 
 }
